# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Eventos",
        "description": "Documentación de la API para la gestión de eventos y reservas.",
        "contact": {
            "email": "tu_correo@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un nuevo usuario en el sistema.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Autentica un usuario existente y retorna un token de acceso.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token de acceso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "Las credenciales son incorrectas"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Cerrar sesión de usuario",
                "description": "Invalida el token de acceso del usuario autenticado.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cierre de sesión exitoso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obtener la lista de eventos",
                "description": "Retorna la lista de eventos disponibles, con soporte para paginación, búsqueda por nombre y filtrado por fecha y ubicación.",
                "operationId": "getEventsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar eventos por fecha (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filtrar eventos por ubicación (parcial o completa)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar eventos por nombre (parcial o completo)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Event"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Crear un nuevo evento",
                "description": "Crea un nuevo evento en el sistema. Requiere autenticación.",
                "operationId": "storeEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del evento",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "date": {
                                        "description": "Fecha del evento (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "location": {
                                        "description": "Ubicación del evento",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripción del evento",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obtener información de un evento específico",
                "description": "Retorna la información de un evento basado en su ID.",
                "operationId": "getEventById",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID del evento a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Actualizar un evento existente",
                "description": "Actualiza la información de un evento existente basado en su ID. Requiere autenticación.",
                "operationId": "updateEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID del evento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Eliminar un evento existente",
                "description": "Elimina un evento existente basado en su ID. Requiere autenticación.",
                "operationId": "deleteEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID del evento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Evento eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{event}/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva para un evento",
                "description": "Crea una nueva reserva para un evento específico. Requiere autenticación.",
                "operationId": "storeReservation",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID del evento para el cual se crea la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "user_email",
                                    "seats"
                                ],
                                "properties": {
                                    "user_name": {
                                        "description": "Nombre del usuario que realiza la reserva",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "user_email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "seats": {
                                        "description": "Número de asientos reservados",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{reservation}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reserva existente",
                "description": "Elimina una reserva existente basada en su ID. Requiere autenticación.",
                "operationId": "deleteReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID de la reserva a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reserva eliminada exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Event": {
                "title": "Event",
                "description": "Modelo de Evento",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID del evento",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Nombre del evento",
                        "type": "string",
                        "example": "Concierto de Rock"
                    },
                    "date": {
                        "title": "Date",
                        "description": "Fecha del evento",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-15"
                    },
                    "location": {
                        "title": "Location",
                        "description": "Ubicación del evento",
                        "type": "string",
                        "example": "Estadio Central"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Descripción del evento",
                        "type": "string",
                        "example": "Un gran concierto de rock en vivo"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Event"
                }
            },
            "Reservation": {
                "title": "Reservation",
                "description": "Reservation model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Reservation ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "event_id": {
                        "title": "Event ID",
                        "description": "ID of the event the reservation belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_name": {
                        "title": "User Name",
                        "description": "Name of the user making the reservation",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "user_email": {
                        "title": "User Email",
                        "description": "Email of the user making the reservation",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "seats": {
                        "title": "Seats",
                        "description": "Number of seats reserved",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Timestamp of when the reservation was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Timestamp of when the reservation was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Reservation"
                }
            }
        }
    },
...
}